+++ app/src/pages/Create/Form/index.js
@@ -1,51 +1,36 @@
 import styled from '@emotion/styled';
-import TextField from '@material-ui/core/TextField';
 import PrimaryButton from '../../../components/atoms/PrimaryButton';
 import { useFormik } from 'formik';
-import { zipObj } from 'ramda';
 import { useEffect } from 'react';
 import CustomizeParameters from './CustomizeParameters';
 import SelectModel from './SelectModel';
-import String from './InputsUI/String';
+import PrimaryInput from './PrimaryInput';
+import { getInitialValues, getInputs } from './utils';
 
-const PROMPT_KEY = 'Prompt';
 
 const Form = ({ 
     onSubmit, isDisabled, models, 
     selectedModel, onSelectModel }) => {
 
-
-    const formik = useFormik({
-        // Initialize the form only with prompt value
-        initialValues: {
-            // Prompt: 'Bauhaus spaceship',
-        },
-        onSubmit,
-        enableReinitialize: true,
-    });
+  const formik = useFormik({
+      initialValues: {},
+      onSubmit,
+      enableReinitialize: true,
+  });
     
-    const inputs = models[selectedModel?.key]?.components.schemas.Input.properties;
-    const primary_input =  inputs ? Object.values(inputs)
-      .map( (values, idx, array) => {
-        const keys = Object.keys(inputs);
-        return {...values, key: keys[idx]}
-      })
-      .sort((a,b) => a['x-order'] > b['x-order'])
-      .shift() : {};
+  const { inputs, primary_input } = getInputs(models, selectedModel);
 
-    useEffect(()=>{
-      // add other fields to the form when user selects the desired model.
-      formik.setValues(state => ({ 
-        ...getInitialValues(inputs, primary_input),
-        [primary_input.key]: formik.values[Object.keys(formik.values)[0]]
-      }))
-    },[selectedModel])
-
-    useEffect(()=>{
-      console.log(formik.values)
-    },[formik.values])
+  useEffect(()=>{
+    // add other fields to the form when user selects the desired model.
+    formik.setValues(state => ({ 
+      // all parameters for the form
+      ...getInitialValues(inputs, primary_input),
+      
+      // override the primary_input value with the old one.
+      [primary_input.key]: state[Object.keys(state)[0]]
+    }))
+  },[selectedModel])
     
-
   return <StyledForm onSubmit={formik.handleSubmit} >
 
     <PrimaryInput
@@ -76,22 +61,6 @@ const Form = ({
 };
 export default Form;
 
-const PrimaryInput = props => {
-
-  const { isDisabled, formik, primary_input } = props;
-
-  return <>
-  <String 
-      value={formik.values[primary_input?.key]} 
-      onChange={formik.handleChange} 
-      name={primary_input.key} 
-      title={primary_input.title} 
-      disabled={isDisabled}
-      fullWidth
-    />
-  
-  </>
-}
 
 const StyledForm = styled.form`
 display: flex;
@@ -99,18 +68,4 @@ flex-direction: column;
 align-items: center;
 gap: 2em;
 width: 100%;
-
 `
-
-function getInitialValues(inputs, primary_input) {
-  if(!inputs) return null
-  const keys = Object.keys(inputs);
-
-  return zipObj( keys, keys?.sort((a,b)=>{
-    console.log(a,b)
-    return inputs[a]['x-order'] > inputs[b]['x-order']
-  }
-    )
-    .map(key => inputs[key]?.default )
-    );
-}
